services/model/pyproject.toml
[project]
name = "model-service"
version = "0.1.0"
description = "Model training and inference service"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.8.2",
    "pyqlib>=0.9.6",
    "torch>=2.3.1",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "yfinance>=0.2.43",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.2",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-p no:cacheprovider"

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.2",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]
